+++ niryo_robot_user_interface/src/niryo_robot_user_interface/tcp_server/command_interpreter.py
@@ -726,6 +726,27 @@ class CommandInterpreter:
         self.__niryo_robot.close_gripper(speed, max_troque_percentage, hold_torque_percentage)
         return self.__send_answer()
 
+    @check_nb_args(4)
+    def __control_gripper(self, position, speed, max_torque, hold_torque):
+        self.__check_type(position, int)
+        self.__check_type(speed, int)
+        self.__check_type(max_torque, int)
+        self.__check_type(hold_torque, int)
+
+        self.__niryo_robot.control_gripper(position, speed, max_torque, hold_torque)
+        return self.__send_answer()
+
+    @check_nb_args(1)
+    def __get_gripper_specs(self, tool_id):
+        tool_id = self.__tools_string_dict_convertor[tool_id]
+        specs = self.__niryo_robot.get_gripper_specs(tool_id)
+        return self.__send_answer(specs)
+
+    @check_nb_args(0)
+    def __get_current_tool_position(self):
+        position = self.__niryo_robot.get_current_tool_position()
+        return self.__send_answer(position)
+
     # - Vacuum
     @check_nb_args(0)
     def __pull_air_vacuum_pump(self):